requirements.txt
mysql-connector-python
argon2-cffi
customtkinter
CTkTable
